Added takexp arguments
Moved around variables and datatypes. Changed the changeExp() in
ExperienceManager to convert to doubles, otherwise would infinite loop
as an int.

+++ src/LevelUP/ExperienceManager.java
@@ -70,7 +70,7 @@ public class ExperienceManager
   
   public void changeExp(int amt)
   {
-    changeExp(amt);
+    changeExp((double)amt);
   }
   
   public void changeExp(double amt)
